+++ History.md
@@ -2,6 +2,7 @@
 ===
 
   * Support `X-Forwarded-Host` in `req.subdomains`
+  * Support IP address host in `req.subdomains`
   * deps: connect@2.26.0
     - deps: body-parser@~1.8.1
     - deps: compression@~1.1.0